cloud_functions/prefetch_schedule/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
nba-api==1.5.*
